PDF: merge overlapping cluster spans where they are consumed
`ShapedRun::cluster_ranges` is the shaper's own segmentation, and every
consumer reads it that way. Only the PDF backend needs the overlapping
clusters folded together for `/ActualText`.
